+++ src/main/java/dnd/studyplanner/config/OAuthService.java
@@ -15,8 +15,8 @@ import dnd.studyplanner.auth.model.AuthEntity;
 import dnd.studyplanner.auth.model.AuthRepository;
 import dnd.studyplanner.config.dto.OAuthAttributes;
 import dnd.studyplanner.jwt.JwtService;
-import dnd.studyplanner.member.model.Member;
-import dnd.studyplanner.member.model.MemberRepository;
+import dnd.studyplanner.domain.member.model.Member;
+import dnd.studyplanner.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
