+++ src/main/java/dnd/studyplanner/repository/MemberRepository.java
@@ -1,8 +1,10 @@
-package dnd.studyplanner.member.model;
+package dnd.studyplanner.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import dnd.studyplanner.domain.member.model.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
