--- src/main/java/dnd/studyplanner/user/model/UserJoinGroup.java
@@ -1,36 +1,0 @@
-package dnd.studyplanner.user.model;
-
-import java.time.LocalDateTime;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-
-import dnd.studyplanner.studygroup.model.StudyGroup;
-import lombok.Getter;
-
-@Getter
-@Entity
-public class UserJoinGroup {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "user_join_group_id")
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
-
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "study_group_id")
-	private StudyGroup studyGroup;
-
-	private LocalDateTime joinDate;
-}
